refactor(navigation): Move navigation into its own file
